pkg/fs: fix test and benchmark
This commit fixes a test and benchmark that I had previously broken
because I hadn't rebased the branch before I submitted it for PR.

Because of d5057b3, which changed a method signature, and because
of a typo, the tests would no longer pass. This did not affect
mainstream code and should only affect developers running on the source.

This commit should fix the issue.