add thumbnail generator for browser
This commit adds a new handler to the minio web server which generates
thumbnails on-the-fly for images (png, gif, and jpeg). It uses the
`nailedit` command, which is a separately packaged binary.

Use is as follows:

GET /minio/thumbnail/{bucket}/{object}?token={token}

Where {bucket} is the bucket name, {object} is the object name, and
{token} is the auth token. It will then load the object from the server,
resize it using `nailedit`, convert it to JPEG and push it to the
response.

User must have an authorized token or have permission for s3:GetObject
to load the thumbnail.