add thumbnail generator for browser
This commit adds a new handler to the minio web server which generates
thumbnails on-the-fly for images (png, gif, and jpeg). It packages the
thumbnail generation in pkg/thumbnail, a standalone library.

Use is as follows:

GET /minio/thumbnail/{bucket}/{object}

Where {bucket} is the bucket name and {object} is the object name.
It will then load the object from the server, resize it using the
thumbnail library, convert it to JPEG and push it to the response.

User must have an authorized token or have permission for s3:GetObject
to load the thumbnail.